pyproject.toml
[project]
name = "maid"
version = "0.0.1"
description = "Personal Maid which does things I tell her to on regular intervals"
authors = [
    {name = "RahulARanger",email = "saihanumarahul66@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "helium (>=5.1.1,<6.0.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.40.12"

